+++ Lib/Core/ProjectManager.h
@@ -80,7 +80,7 @@ namespace UI {
         void removeTabTriggered(const QString& fileName);
         void goToStartWindowTriggered();
 
-      void applySettings(int value1, int value2, int value3);
+      void applySettings(double value1, double value2, double value3);
 
     private:
         void connectStartWindowSignals();
